adicionado funcao pra verificar existencia de arquivos antes de remover

+++ nvim/setup_nvim.sh
@@ -4,8 +4,11 @@
 sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
 
-rm $HOME/.config/nvim/init.vim
-ln $PWD/init.vim $HOME/.config/nvim/init.vim || cp $PWD/init.vim $HOME/.config/nvim
+if test -f "$HOME/.config/nvim/init.vim"; then
+  rm $HOME/.config/nvim/init.vim
+fi
+
+ln init.vim $HOME/.config/nvim/init.vim || cp init.vim $HOME/.config/nvim
 
 # install coc.vim extension
 nvim +PlugInstall +qa
